Move DPOSv2 Oracle from loomchain
