+++ Makefile
@@ -27,7 +27,7 @@ TRANSFER_GATEWAY_DIR=$(GOPATH)/src/$(PKG_TRANSFER_GATEWAY)
 #       specific commit.
 GO_LOOM_GIT_REV = HEAD
 # Specifies the loomnetwork/transfer-gateway branch/revision to use.
-TG_GIT_REV = HEAD
+TG_GIT_REV = move-dpos-oracle
 # loomnetwork/go-ethereum loomchain branch
 ETHEREUM_GIT_REV = 1fb6138d017a4309105d91f187c126cf979c93f9
 # use go-plugin we get 'timeout waiting for connection info' error
@@ -102,8 +102,8 @@ tron_tgoracle: $(TRANSFER_GATEWAY_DIR)
 pcoracle:
 	go build $(GOFLAGS) -o $@ $(PKG)/cmd/$@
 
-dposv2_oracle:
-	go build $(GOFLAGS) -o $@ $(PKG)/cmd/$@
+dposv2_oracle: $(TRANSFER_GATEWAY_DIR)
+	go build $(GOFLAGS_GATEWAY) -o $@ $(PKG_TRANSFER_GATEWAY)/cmd/$@
 
 loom: proto
 	go build $(GOFLAGS) $(PKG)/cmd/$@
