+++ cmd/loom/loom.go
@@ -31,8 +31,6 @@ import (
 	"github.com/loomnetwork/loomchain/abci/backend"
 	"github.com/loomnetwork/loomchain/auth"
 	"github.com/loomnetwork/loomchain/builtin/plugins/dposv2"
-	d2Oracle "github.com/loomnetwork/loomchain/builtin/plugins/dposv2/oracle"
-	d2OracleCfg "github.com/loomnetwork/loomchain/builtin/plugins/dposv2/oracle/config"
 	"github.com/loomnetwork/loomchain/builtin/plugins/dposv3"
 	plasmaConfig "github.com/loomnetwork/loomchain/builtin/plugins/plasma_cash/config"
 	plasmaOracle "github.com/loomnetwork/loomchain/builtin/plugins/plasma_cash/oracle"
@@ -425,10 +423,6 @@ func newRunCommand() *cobra.Command {
 				return err
 			}
 
-			if err := startDPOSv2Oracle(chainID, cfg.DPOSv2OracleConfig); err != nil {
-				return err
-			}
-
 			if err := startFeatureAutoEnabler(chainID, cfg.ChainConfig, nodeSigner, backend, log.Default); err != nil {
 				return err
 			}
@@ -451,25 +445,6 @@ func recovery() {
 		log.Error("caught RPC proxy exception, exiting", r)
 		os.Exit(1)
 	}
-}
-
-func startDPOSv2Oracle(chainID string, cfg *d2OracleCfg.OracleSerializableConfig) error {
-	oracleCfg, err := d2OracleCfg.LoadSerializableConfig(chainID, cfg)
-	if err != nil {
-		return err
-	}
-
-	if !oracleCfg.Enabled {
-		return nil
-	}
-
-	oracle := d2Oracle.NewOracle(oracleCfg)
-	if err := oracle.Init(); err != nil {
-		return err
-	}
-
-	oracle.Run()
-	return nil
 }
 
 func startFeatureAutoEnabler(
